// backend/src/models/TvshowModel.ts
export class TvShowModel{
    title: string = '';
    release_date: string = '';
    ending_date: string = '';
    number_of_episodes: string = '';
    number_of_seasons: string = '';
    overview: string = '';
    language: string = '';
    status: string = '';
    rate: number = 0;
    genre: [] = [];
    poster: string = 'https://image.tmdb.org/t/p/original';
    bg_poster: string = 'https://image.tmdb.org/t/p/original';
}