// backend/src/models/MovieModel.ts
export class MovieModel{
    title: string = '';
    release: string = '';
    overview: string = '';
    language: string = '';
    status: string = '';
    rate: number = 0;
    genre: [] = [];
    poster: string = 'https://image.tmdb.org/t/p/original';
    bg_poster: string = 'https://image.tmdb.org/t/p/original';
}