// backend/src/models/AnimationModel.ts
export class AnimationModel{
    title:  string = '';
    overview: string = '';
    release_date: string = '';
    language: string = '';
    rate: number = 0;
    genre: [] = [];
    poster: string = 'https://image.tmdb.org/t/p/original';
    bg_poster: string = 'https://image.tmdb.org/t/p/original'
}